Dev (#18)
* Agregando peluqueros en la fake data

* Fix buscar peluqueros (#16)

* se agrega request de buscar que recibe una ubicacion, filtros por nombre, tipo y tipo de servicio (opcionales) y un paginable (opcional) y devuelve una pagina de peluqueros

* se cambia nombre de PeluqueroType CHICOS a KIDS y se agrega UNISEX

* filtro de tipo de peluquero permite elegir multiples por logica OR y se elimina tipo UNISEX

* borrado vieja implementacion de buscar peluqueros y funciones sin usar

* Pedir turnos (#17)

* Agregando emailSender para el envio de emails

* Agregando un threadpool aparte para ejecutar el envio de emails en distintos threads

* Agrego clase Turno con sus componentes y Cliente. Agrego tests de comportamiento sobre Turno.

* Finalizo la parte del back con los endpoints para pedir y confirmar turno (con el envio de email en este ultimo)

* Modifico query turnoRepository que tenia bug. Ahora el confirmar turno setea al peluquero en OCUPADO. Agrego endpoint finalizar turno y cuando el peluquero no tenga turnos confirmados, este pasa a ser DISPONIBLE de nuevo.

* Agrego endpoint para getear a peluquero por id

* Agrego requestbody annotation en los Post y valid en get peluquero por id

* El precio del turno tiene como base siempre el corteMin. Fix tests turno precio

* Update README.md como instalar y ejecutar el backend

+++ src/main/kotlin/ar/edu/unq/peluqueriayabackend/persistence/impl/repositories/ServicioRepository.kt
@@ -2,14 +2,11 @@ package ar.edu.unq.peluqueriayabackend.persistence.impl.repositories
 
 import ar.edu.unq.peluqueriayabackend.model.Peluquero
 import ar.edu.unq.peluqueriayabackend.model.Servicio
-import ar.edu.unq.peluqueriayabackend.model.ServicioType
-import org.springframework.data.domain.Page
 import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ServicioRepository : JpaRepository<Servicio, Int> {
+interface ServicioRepository : JpaRepository<Servicio, Long> {
 
     fun findAllByPeluquero(peluquero: Peluquero): Collection<Servicio>
 }
